ResolvedLicenseInfo: add applyChoices function for license choices
For later use.
